debian/: Update deps for Debian 10 & Raspbian 10
I think the gir1.2-vte dependency is not strictly correct (it's gtk3
only) but this dependency will be part of the way forward for gscreen's
terminal widget when it transitions to gtk3.

This yields a working package which can be installed, and AXIS runs.

(cherry picked from commit e31ba3c8b8a39284d98413b71a7f677bd8927d54)